+++ src/generator/grid/mod.rs
@@ -1,5 +1,4 @@
 pub mod dungeon;
 pub mod faction;
-pub mod faction2;
 pub mod maze;
 pub mod room;
