+++ src/generator/mod.rs
@@ -3,8 +3,6 @@ mod unaligned;
 
 pub use self::grid::dungeon::DungeonGen;
 pub use self::grid::faction::FactionGen;
-pub use self::grid::faction2::inner_tick;
-pub use self::grid::faction2::FactionGen as FactionGen2;
 pub use self::grid::maze::MazeGen;
 pub use self::grid::room::RoomGen;
